requirements.txt
# ── Base de datos ────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
pyodbc==5.1.0

# ── Data ─────────────────────────────────────────────────────────────────────
pandas==2.2.3

# ── Excel / CSV ───────────────────────────────────────────────────────────────
openpyxl==3.1.5
xlsxwriter==3.2.0

# ── Dashboards ───────────────────────────────────────────────────────────────
dash==2.18.1
plotly==5.24.1

# ── Utilidades ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tabulate==0.9.0
